Move ticket forms into a new module so I can have both open at once
